Dynamic footer menu
When the window width goes below 750px (mobile size), the menu with more
information at the bottom of home.html will now transform into a menu of
the same style as the navigation bar on all the other pages. "Enrolling
for courses/exams" has been renamed to "Enrollment" to reduce the amount
of space required.