Mobile blocks
Two rows of two panels have now been added for the mobile version. The
background of these blocks is the picture that would normally be in the
body of the panel. The text in the headings and the footers will jump to
the next line at the exact same moment.